requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.7